chore: remove event_log table
